+++ MEGPAY.Account/Account.Domain/AccountAggregate/Account.cs
@@ -6,9 +6,9 @@ public class Account : Entity<Guid>
 {
     public string Name { get; set; }
     public string Surname { get; set; }
-    public double CurrentBalance { get; set; }
+    public decimal CurrentBalance { get; set; }
 
-    public Account(Guid id, string name, string surname, double currentBalance)
+    public Account(Guid id, string name, string surname, decimal currentBalance)
     {
         Id = id;
         Name = name;
@@ -16,7 +16,7 @@ public class Account : Entity<Guid>
         CurrentBalance = currentBalance;
     }
 
-    public void UpdateBalance(double currentbalance)
+    public void UpdateBalance(decimal currentbalance)
     {
         CurrentBalance = currentbalance;
     }
